+++ backend/src/main/java/br/edu/ifs/playifs/services/CourseService.java
@@ -5,38 +5,34 @@ import br.edu.ifs.playifs.entities.Campus;
 import br.edu.ifs.playifs.entities.Course;
 import br.edu.ifs.playifs.repositories.CampusRepository;
 import br.edu.ifs.playifs.repositories.CourseRepository;
-import br.edu.ifs.playifs.services.exceptions.DatabaseException;
+import br.edu.ifs.playifs.services.exceptions.BusinessException;
 import br.edu.ifs.playifs.services.exceptions.ResourceNotFoundException;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 public class CourseService {
 
     @Autowired
     private CourseRepository repository;
-
     @Autowired
-    private CampusRepository campusRepository; // Injetamos o repositório de Campus
+    private CampusRepository campusRepository;
 
     @Transactional(readOnly = true)
     public CourseDTO findById(Long id) {
-        Course course = repository.findById(id).orElseThrow(
-                () -> new ResourceNotFoundException("Recurso não encontrado"));
-        return new CourseDTO(course);
+        Course entity = repository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Curso não encontrado com o ID: " + id));
+        return new CourseDTO(entity);
     }
 
     @Transactional(readOnly = true)
-    public List<CourseDTO> findAll() {
-        List<Course> list = repository.findAll();
-        return list.stream().map(CourseDTO::new).collect(Collectors.toList());
+    public Page<CourseDTO> findAll(String name, Long campusId, Pageable pageable) {
+        Page<Course> page = repository.find(name, campusId, pageable);
+        return page.map(CourseDTO::new);
     }
 
     @Transactional
@@ -55,28 +51,25 @@ public class CourseService {
             entity = repository.save(entity);
             return new CourseDTO(entity);
         } catch (EntityNotFoundException e) {
-            throw new ResourceNotFoundException("Recurso não encontrado");
+            throw new ResourceNotFoundException("Recurso não encontrado com o ID: " + id);
         }
     }
 
-    @Transactional(propagation = Propagation.SUPPORTS)
+    @Transactional
     public void delete(Long id) {
         if (!repository.existsById(id)) {
-            throw new ResourceNotFoundException("Recurso não encontrado");
+            throw new ResourceNotFoundException("Recurso não encontrado com o ID: " + id);
         }
-        try {
-            repository.deleteById(id);
-        } catch (DataIntegrityViolationException e) {
-            throw new DatabaseException("Falha de integridade referencial");
+        Course course = repository.findById(id).get();
+        if (!course.getTeams().isEmpty()) {
+            throw new BusinessException("Não é possível apagar um curso que já possui equipas inscritas.");
         }
+        repository.deleteById(id);
     }
 
-    // Método auxiliar para copiar os dados do DTO para a entidade
     private void copyDtoToEntity(CourseDTO dto, Course entity) {
         entity.setName(dto.getName());
         entity.setLevel(dto.getLevel());
-
-        // Busca a entidade Campus pelo ID fornecido no DTO e a associa ao curso
         Campus campus = campusRepository.getReferenceById(dto.getCampus().getId());
         entity.setCampus(campus);
     }
